LayoutTests/imported/w3c/web-platform-tests/svg/shapes/animatedPoints-non-animated-expected.txt


FAIL animatedPoints represents non-animated value assert_equals: expected 20 but got 170
